Remove old loader, rename loader_simple to loader
