+++ applications/services/loader/application.fam
@@ -4,22 +4,8 @@ App(
     apptype=FlipperAppType.SERVICE,
     entry_point="loader_srv",
     cdefines=["SRV_LOADER"],
-    requires=["gui"],
-    provides=["loader_start"],
     stack_size=2 * 1024,
     order=90,
-    targets=["f7", "f18"],
-    sdk_headers=[
-        "loader.h",
-        "firmware_api/firmware_api.h",
-    ],
-)
-
-App(
-    appid="loader_start",
-    apptype=FlipperAppType.STARTUP,
-    entry_point="loader_on_system_start",
-    requires=["loader"],
-    order=90,
-    targets=["f7", "f18"],
+    targets=["f20"],
+    sources=["loader.c"],
 )
