+++ applications/services/loader/loader_i.h
@@ -1,38 +1,26 @@
 #pragma once
 #include <furi.h>
 #include <toolbox/api_lock.h>
-#include <flipper_application/flipper_application.h>
 #include "loader.h"
-#include "loader_menu.h"
-#include "loader_applications.h"
 
 typedef struct {
     char* args;
     FuriThread* thread;
     bool insomniac;
-    FlipperApplication* fap;
 } LoaderAppData;
 
 struct Loader {
     FuriPubSub* pubsub;
     FuriMessageQueue* queue;
-    LoaderMenu* loader_menu;
-    LoaderApplications* loader_applications;
     LoaderAppData app;
 };
 
 typedef enum {
     LoaderMessageTypeStartByName,
     LoaderMessageTypeAppClosed,
-    LoaderMessageTypeShowMenu,
-    LoaderMessageTypeMenuClosed,
-    LoaderMessageTypeApplicationsClosed,
     LoaderMessageTypeLock,
     LoaderMessageTypeUnlock,
     LoaderMessageTypeIsLocked,
-    LoaderMessageTypeStartByNameDetachedWithGuiError,
-    LoaderMessageTypeSignal,
-    LoaderMessageTypeGetApplicationName,
 } LoaderMessageType;
 
 typedef struct {
@@ -42,24 +30,7 @@ typedef struct {
 } LoaderMessageStartByName;
 
 typedef struct {
-    uint32_t signal;
-    void* arg;
-} LoaderMessageSignal;
-
-typedef enum {
-    LoaderStatusErrorUnknown,
-    LoaderStatusErrorInvalidFile,
-    LoaderStatusErrorInvalidManifest,
-    LoaderStatusErrorMissingImports,
-    LoaderStatusErrorHWMismatch,
-    LoaderStatusErrorOutdatedApp,
-    LoaderStatusErrorOutOfMemory,
-    LoaderStatusErrorOutdatedFirmware,
-} LoaderStatusError;
-
-typedef struct {
     LoaderStatus value;
-    LoaderStatusError error;
 } LoaderMessageLoaderStatusResult;
 
 typedef struct {
@@ -72,8 +43,6 @@ typedef struct {
 
     union {
         LoaderMessageStartByName start;
-        LoaderMessageSignal signal;
-        FuriString* application_name;
     };
 
     union {
