--- applications/services/loader_simple/application.fam
@@ -1,15 +1,0 @@
-App(
-    appid="loader_simple",
-    name="LoaderSrv",
-    apptype=FlipperAppType.SERVICE,
-    entry_point="loader_srv",
-    cdefines=["SRV_LOADER"],
-    requires=["gui"],
-    stack_size=2 * 1024,
-    order=90,
-    targets=["f50", "f20"],
-    sdk_headers=[
-        # "loader.h",
-        # "firmware_api/firmware_api.h",
-    ],
-)
