--- applications/services/loader_simple/firmware_api/firmware_api.h
@@ -1,5 +1,0 @@
-#pragma once
-
-#include <flipper_application/elf/elf_api_interface.h>
-
-extern const ElfApiInterface* const firmware_api_interface;
