--- applications/services/loader_simple/loader.h
@@ -1,70 +1,0 @@
-#pragma once
-#include <furi.h>
-
-#ifdef __cplusplus
-extern "C" {
-#endif
-
-#define RECORD_LOADER            "loader"
-#define LOADER_APPLICATIONS_NAME "Apps"
-
-typedef struct Loader Loader;
-
-typedef enum {
-    LoaderStatusOk,
-    LoaderStatusErrorAppStarted,
-    LoaderStatusErrorUnknownApp,
-    LoaderStatusErrorInternal,
-} LoaderStatus;
-
-typedef enum {
-    LoaderEventTypeApplicationBeforeLoad,
-    LoaderEventTypeApplicationLoadFailed,
-    LoaderEventTypeApplicationStopped
-} LoaderEventType;
-
-typedef struct {
-    LoaderEventType type;
-} LoaderEvent;
-
-/**
- * @brief Start application
- * @param[in] instance loader instance
- * @param[in] name application name or id
- * @param[in] args application arguments
- * @param[out] error_message detailed error message, can be NULL
- * @return LoaderStatus
- */
-LoaderStatus
-    loader_start(Loader* instance, const char* name, const char* args, FuriString* error_message);
-
-/**
- * @brief Lock application start
- * @param[in] instance loader instance
- * @return true on success
- */
-bool loader_lock(Loader* instance);
-
-/**
- * @brief Unlock application start
- * @param[in] instance loader instance
- */
-void loader_unlock(Loader* instance);
-
-/**
- * @brief Check if loader is locked
- * @param[in] instance loader instance
- * @return true if locked
- */
-bool loader_is_locked(Loader* instance);
-
-/**
- * @brief Get loader pubsub
- * @param[in] instance loader instance
- * @return FuriPubSub* 
- */
-FuriPubSub* loader_get_pubsub(Loader* instance);
-
-#ifdef __cplusplus
-}
-#endif
