--- applications/services/loader_simple/loader_i.h
@@ -1,52 +1,0 @@
-#pragma once
-#include <furi.h>
-#include <toolbox/api_lock.h>
-#include "loader.h"
-
-typedef struct {
-    char* args;
-    FuriThread* thread;
-    bool insomniac;
-} LoaderAppData;
-
-struct Loader {
-    FuriPubSub* pubsub;
-    FuriMessageQueue* queue;
-    LoaderAppData app;
-};
-
-typedef enum {
-    LoaderMessageTypeStartByName,
-    LoaderMessageTypeAppClosed,
-    LoaderMessageTypeLock,
-    LoaderMessageTypeUnlock,
-    LoaderMessageTypeIsLocked,
-} LoaderMessageType;
-
-typedef struct {
-    const char* name;
-    const char* args;
-    FuriString* error_message;
-} LoaderMessageStartByName;
-
-typedef struct {
-    LoaderStatus value;
-} LoaderMessageLoaderStatusResult;
-
-typedef struct {
-    bool value;
-} LoaderMessageBoolResult;
-
-typedef struct {
-    FuriApiLock api_lock;
-    LoaderMessageType type;
-
-    union {
-        LoaderMessageStartByName start;
-    };
-
-    union {
-        LoaderMessageLoaderStatusResult* status_value;
-        LoaderMessageBoolResult* bool_value;
-    };
-} LoaderMessage;
